requirements.txt
certifi==2019.6.16
chardet==3.0.4
distro==1.4.0
idna==2.8
numpy==1.16.4
pandas==0.24.2
pdf2textbox==0.4.3
pdfminer2==20151206
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
tabula-py==1.3.1
urllib3==1.25.3
